refactor: project directory
