--- src/app/hyuabot/api/api/fetch/__init__.py
@@ -1,3 +1,0 @@
-from fastapi import APIRouter
-
-fetch_router = APIRouter(prefix="/fetch")
