+++ src/app/hyuabot/api/core/fetch/__init__.py
@@ -1,0 +1,3 @@
+from fastapi import APIRouter
+
+fetch_router = APIRouter(prefix="/fetch")
