+++ src/app/hyuabot/api/main.py
@@ -1,7 +1,7 @@
 from fastapi import FastAPI
 
 from app.hyuabot.api.api.api_v1.api import shuttle_router, bus_router
-from app.hyuabot.api.api.fetch import fetch_router
+from app.hyuabot.api.core.fetch import fetch_router
 from app.hyuabot.api.core.config import settings
 from app.hyuabot.api.initialize_data import initialize_data
